pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "architecture-control-plane"
version = "0.1.0"
description = "Control plane scaffold for cycle orchestration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "httpx>=0.27.0",
  "PyJWT>=2.8.0",
  "prometheus-client>=0.20.0",
  "cryptography>=42.0.0",
  "jsonschema>=4.23.0",
  "openai>=1.99.0",
  "PyYAML>=6.0.2",
  "opentelemetry-api>=1.28.0",
  "opentelemetry-sdk>=1.28.0",
  "opentelemetry-exporter-otlp-proto-http>=1.28.0",
  "boto3>=1.35.0",
  "redis>=5.0.0",
]


[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
  "postgres_integration: requires a reachable PostgreSQL database",
]


[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["db_migrations*", "tests*"]


[tool.ruff.lint]
ignore = ["F403", "F405"]
